Fix issues

+++ src/main/java/com/nitesh/jukebox/dao/scylla/ArtistDao.java
@@ -4,9 +4,7 @@ package com.nitesh.jukebox.dao.scylla;
 import com.datastax.driver.mapping.Mapper;
 import com.nitesh.jukebox.dao.resource.ScyllaResource;
 import com.nitesh.jukebox.dao.scylla.accessor.JukeBoxArtistAccessor;
-import com.nitesh.jukebox.dao.scylla.accessor.JukeBoxMovieAccessor;
 import com.nitesh.jukebox.models.entity.Artist;
-import com.nitesh.jukebox.models.entity.Movie;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Repository;
@@ -46,7 +44,7 @@ public class ArtistDao {
         artistMapper.delete(artistId);
     }
 
-    public List<Artist> getArtistsByRating(final int low, final int high) {
+    public List<Artist> getArtistsByRating(final double low, final double high) {
         JukeBoxArtistAccessor accessor = scyllaResource.getAccessor(JukeBoxArtistAccessor.class);
         return accessor.getArtistsByRating(low, high).all();
     }
